fusedchat_backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database - MongoDB
pymongo==4.6.0
motor==3.3.2

# LLM & RAG
langchain-community==0.0.13
langchain-huggingface==0.0.8
langchain-text-splitters==0.0.1
langchain-core==0.1.33

# Vector Database
faiss-cpu==1.7.4
chromadb==0.4.14

# Document Processing
fitz==0.0.1
PyMuPDF==1.23.8
pypdf==4.0.1
python-dotenv==1.0.0
tiktoken==0.5.2
rapidocr-onnxruntime==1.3.13

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0


# Utilities
requests==2.31.0
aiohttp==3.9.1